Added C4DtoA Obq_StereoLens UI
Added a new Cinema4D C4DtoA Obq_StereoLens customized UI and updated the
Obq_AngularCamera UI.

Note: There will need to be more error handling
added to the main Obq_StereoLens shader source code to better handle an
invalid/missing camera names in the shader when a new C4DtoA render is
launched to avoid crashes.

Also C4DtoA can't define a custom enum menu
if the data type of the variable is an int. This causes the
Obq_StereoLens "View Mode" attribute to display as a number field in the
Cinema4D UI.